backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
auditlog3==1.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pluggy==0.13.1
psycopg2-binary==2.9.7
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sorl-thumbnail==12.8.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
